+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-### 2025-05-16T21:20:44+02:00
-- Originele XML-bestanden hersteld vanuit v19A.1.8 ZIP zonder wijzigingen.
+### 2025-05-16T21:43:29+02:00
+- MainActivity.kt gecorrigeerd: vier bestaande knoppen gekoppeld aan CallActivity, ChatListActivity, LanguageSelectionActivity en ThemeSettingsActivity.